Get nodeid through hostname and add Jmeter support
MUSIC 1.0.1: This commit has two independent changes: (i) Changed MusicUtil.java to generate the
node id through the hostname rather than host IP address (ii) Added a class MusicJmeterClient that
has a list of all tests that a Jmeter client may want to run.

The first change was in response to the fact that sometimes different nodes generate the same id
when we use their local host IP adress (that of 127.0.0.1) while the expectation was that they will
always generate the VM private IP. So using the hostname to generate id which has a better chance of
being unique across nodes.

The second change was simply to enable Jmeter testing wherein the tests can simply call each of the
8-9 potential functions such as cassaGet, CassaPut, zkGet and so on.