1.0.7: Add music.properties, change log4j location
The older versions of MUSIC were using machine host names for generating unique node ids. Changing
that to obtain it explicitly from a properties file that needs to be placed in
/etc/music/music.properties.

Moving location of log4j.properties from WEB-INF/ to /etc/music/ since in the former location it
will always be overwritten by the unpacking of a MUSIC.war file.