Hal -v 1.0.1: Address core restart issue
This fix tried to address the CQOS-77 issue wherein the active hal would keep trying to start a core
service in case it comes up successfully and then goes down.

The fix was to associate a configurable retry attempt that is decremented every time the core has to
be restarted. Once the attempts are exhausted, the active cedes control and becomes passive.