Add new jmeter jar (1.2.0) to allow atomic put
In the latest commit to MUSIC, version 2.0.0, we added atomic puts wherein MUSIC
acquires the lock to a key, performs an insert/update/delete and then releases the key. To exercise
this command in the benchmarks, have added an operation type called "music_atomic_put" to the
musicJmeter.jar. To exercise this function, in the op_type of musicBm.properties, specify
"music_atomic_put". Note that the latest version of MUSIC.war, musicJmeter.jar and possibly the
musicJmeter_lib need to be updated for this.