MUSIC 2.1.0: Refactor Code
Mainly refactored the rest API to help in understanding. One important change is that in the MUSIC q
api, now the URL to be used is /priorityq/. So for example, the path:

/keyspaces/{keyspace}/priorityq/{qname} will now change to /priorityq/keyspaces/{keyspace}/{qname}

All of this functionality is in RestMusicQAPI.

Redocumenting the API change of MUSIC 2.0.0 for convenience: A crucial API change is that, in an insert/update/delete previously the consistency option of
"atomic" is now called "critical" wherein the client has to acquire the lock first. The "atomic"
option now refers to the atomic put where MUSIC acquires the lock, performs the insert/update/delete
and releases the lock. In summary with any insert/update/delete operation there are three
consistency options that can be specified in the json body:

1. "consistencyInfo": {
        "type": "eventual"
    }

In this option it is a simple eventual put.

2. "consistencyInfo": {
        "type": "atomic"
    }

In this option it is an atomic put wherein MUSIC acquires the lock, performs the write and then
releases the lock.

3. "consistencyInfo": {
        "type": "critical"
        "lockId": "<whatever the lock id is>"
    }

In this option, the client has to acquire the lock and provide a lockId to MUSIC. If the client is
the lockholder MUSIC will perform the operation.