MUSIC 2.2.0: Add atomic test and set
Extended the update/delete API to support atomic test condition and update/delete the row if the
condition is true. MUSIC will return true if it is able to acquire the lock and the condition is
true and the update is successful.

So if the type of update/delete is "atomic" or "critical" you can also add an optional "conditions" json field that
has column name and column value pairs. MUSIC will first acquire the lock to the row specified, read
the row (from a quorum of replicas) and if its columns match the conditions, perform the
update/delete on a quorum of replicas. For example, the json
body for an update could be like this:
{
    "conditions":{
    	"emp_salary":4000
    },
    "values": {
        "emp_salary": 5000
    },
    "consistencyInfo": {
        "type": "atomic"
    }
}

Here, MUSIC will atomically update the salary of the row if its salary column contains 4000.
Currently the conditions supported are only for basic data types and not for list and map.