MUSIC 2.2.5 Fix bugs in new atomic put
This releases fixes two important bugs in the conditional atomic put (CAP)

1. When a CAP is attempted and it fails because it could not acquire the lock, a lock reference is
created that was not previously deleted. This can cause a situation where an infinite loop of
attempted CAP creates infinite lock references. Fixed this by explicitly deleting the lock
reference.

2. Better exception handling to ensure readable error messages when any part of the CAP fails. For
example, if there are rows in the table and Cassandra throws and exception while checking for the
condition, I return this exception to the client.