MUSIC 2.2.11 Add atomic_delete_lock option
This commit has two changes both mainly for the benchmarking.

(1) Along with the "eventual", "critical" and "atomic" options in the consistency info of
the MUSIC update operation we also now provide "atomic_delete_lock" in which the sequence of
operations is:

create lock;
acquire lock;
check condition;
perform update;
delete lock;

as opposed to releasing the lock in the "atomic" option. Example usage:

{
    "conditions":{
    	"emp_salary": 200008
    },
    "values": {
        "emp_salary": 300008
    },
    "consistencyInfo": {
        "type": "atomic_delete_lock"
    }
}

The same consistency option will also work for the zookeeper atomic
update. Hence now for both the zk atomic put and the MUSIC atomic put
by using this option we can differentiate between releasing a lock
and completely deleting it.

(2) In the previous zk atomic put operation there was bug in that
all of them were modifying a single bmObject which might have penalized
zk atomic put performance unfairly. In this version, we use a new URL for
the zk atomic put:

http://localhost:8080/MUSIC/rest/benchmarks/purezk/atomic/{lockname}

For example, the lockname could be testks.employees.bharath indicating that
it is a lock to the keyspace testks, table employees and row primary id bharath.

Now in a zk atomic put rather than update a separate bmObject, we simply update the
lockname object in zookeeper (which will be different for each thread. This will
make it a fair comparison with the MUSIC atomic put in which we update a row in Cassandra.