MUSIC 2.2.17 Add batchsize option for benchmarking
For the zk atomic put and music atomic put you can now add batchsize as a json parameter to test how the critical sections perform for large batch sizes. This is only for benchmarking purposes and should be used only for the zk atomic put and music atomic put (release option for both).

Example json for 500 opertaions:

{
    "values": {
        "emp_salary": 4000
    },
    "consistencyInfo": {
        "type": "atomic"
    },
    "batchSize":500
}

If this parameter is left empty it will default to 1.