MUSIC 2.2.19: Add mix functions for benchmarking
Introduced two new functions:

/MUSIC/rest/benchmarks/zk_mix/keyspaces/{keyspacename}/tables/{tablename}
/MUSIC/rest/benchmarks/music_mix/keyspaces/{keyspacename}/tables/{tablename}

that execute operations mixing between atomic and eventual operation. The json input for both are same and here is the example:

{
   "baseKeyValue": "bharath",
   "startId":1,
   "mixSize":10,
   "keyName":"emp_name",
   "evAtRatio":3,
    "values": {
        "emp_salary": 1000
    }
}

Let us assume this is used for the music_mix. MUSIC generates a range of ids between the baseKeyValue+startId (in this example: bharath1 to bharath10) and executes critical or eventual operations based on whether the number part of the id is divisible by the evAtRatio. Clearly this is only for benchmarking and cannot be generalized if the key format changes. So in this example, MUSIC will execute critical operations for bharath3, bharath6, bharath9 and eventual for remaining. So this will ensure that only 1/3rd of all operations are critical. If the evAtRatio is set to 4, then 1/4th of all operations will be critical and so on. The value field is standard json update that tells you which column to update.