MUSIC 2.2.22 Add support for data size
Introducing a new function with the following example URL:

http://localhost:8080/MUSIC/rest/benchmarks/gen_mix/keyspaces/bmkeyspace/tables/bmemployees

with example json body:

{
   "keyName":"name",
   "baseKeyValue": "emp",
   "startId":0,
   "mixSize":1,
   "evAtRatio":1,
    "values": {
        "job": "loafer"
    },
    "operationType":"Music",
    "dataSize":30
}

The fields are mostly the same as the zk and music mix functions from MUSIC 2.2.20 except for the following:

operationType: use this to combine both music and zk operations into one URL. Specify "Music" or "Zk" (case sensitive) based on the type.

dataSize: if the number here is x, then I create a string of size x*1000 full of "*" and update the "job" field of the specified employee with that string. Hence this is completely hard coded for our benchmarks.

Experiment to be performed:

- Fix mix size at 1, evAtRatio:1 (always atomic)
- Consider three data sizes, i.e. the size of the json body we are sending in our update 1, 512, 1000
- Two usual latency profiles l2 and l3.
- Compare music_atomic_put_with_release and zk_atomic_put_with_release for the usual three trials.