MUSIC 2.2.25: Music version for size experiments
http://localhost:8080/MUSIC/rest/benchmarks/gen_mix/keyspaces/bmkeyspace/tables/bmemployees

with example json body:

{
   "keyName":"name",
   "baseKeyValue": "emp",
   "startId":0,
   "mixSize":1,
   "evAtRatio":1,
    "values": {
        "job": "loafer"
    },
    "operationType":"Music",
    "dataSize":30
}

The fields are mostly the same as the zk and music mix functions from MUSIC 2.2.20 except for the following:

operationType: use this to combine both music and zk operations into one URL. Specify "Music" or "Zk" or  "MusicPureCassa" (case sensitive) based on the type.

dataSize: if the number here is x, then I create a string of size x*1000 full of "*" and update the "job" field of the specified employee with that string. Hence this is completely hard coded for our benchmarks.

Experiment to be performed:

- Fix mix size at 1, evAtRatio:1 (always atomic)
- Consider three data sizes, i.e. the size of the json body we are sending in our update 1, 64, 256
- Two usual latency profiles l2 and l3.

More info on the new operation:

Built a new locking service entirely based on Cassandra's light-weight transactions.

To benchmark it, first call the following API:

http://localhost:8080/MUSIC/rest/benchmarks/locktable/keyspaces/bmkeyspace/tables/bmemployees

to create a locking table (no need to populate it with anything) and then,

in the gen_mix URL:

http://localhost:8080/MUSIC/rest/benchmarks/gen_mix/keyspaces/bmkeyspace/tables/bmemployees

in the json body (e.g):

{
   "keyName":"name",
   "baseKeyValue": "emp",
   "startId":0,
   "mixSize":4,
   "evAtRatio":2,
    "values": {
        "job": "goafer"
    },
    "operationType":"Music",
    "dataSize":1
}

in operationType use "MusicPureCassa".